Fixed issue with tasks runing into a timeout if submitted after cancelling a batch task.
Currently all SolverTask instances try to acquire a lock on the class,
to ensure that only one is running at a time.  Once the lock is acquired
the actual task is submitted to ProB. If a task was cancelled while
waiting for the lock it might acquire the lock before realizing that the
task was actually cancelled, thus Submitting the task to ProB and then
exiting. This behaviour would leave ProB blocked while it finishes the
task in the background making all other tasks wait and eventually
timeout trying to submit tasks to ProB.

Current solution is to check for task cancellation when entering the
critical section, before submitting tasks to ProB.