PDF rendering tasks now have a shared single threaded executor
Custom executor for pdf rendering tasks is used to submit solver tasks.
The goal is to run tasks in separate queues to avoid starvation in case
all threads are running pdf rendering tasks and there are no threads to
run the solver tasks required to generate the pdf.

Additionally updated title and some of the messages for the task

Closes #141