Updated MapChangeListener used in ConflictMatrix.
Only reset the state of the ResultCell associated to the key in the 
change event instead of resetting everything and redrawing the 
matrix for each key change.

The previous behaviour would cause the application to freeze if the cache 
already contained many values since it would redraw everything for each 
key changed.