Refactored ProbSolver in order to drop the PowerMock dependency
Original problem was that the GetOperationByPredicateCommand class
is final and thus cannot be easily mocked using Mockito.

The previous solution was to use PowerMock to mock the final class and to
intercep instantiations.

The new approach uses a delegate object for GetOperationByPredicateCommand.
ProBSolver now takes a factory used to create instances of the delegate.

The default factory/delegate just create a GetOperationByPredicateCommand while
in for testing the factory and command delegate can by mocked.